Refactor profile components and improve error handling
- Updated error handling in BasicInfoForm, PasswordInfoForm, and useFollow hook to provide more descriptive error messages.
- Enhanced type definitions for MediaCard, MediaGrid, ProfileCard, and ProfileGrid components to improve type safety.
- Refactored ProfileHeader and ProfilePage components to use TypeScript interfaces for props.
- Added ApiResponse and FetchDataType types to api.types.ts for better API response handling.
- Improved error handling in EditAlbum, EditPhoto, NewAlbum, and NewPhoto pages to handle Axios errors more gracefully.
- Updated the tsconfig.json to include an output directory for compiled files.
- Added new types for media and profile in their respective type definition files.
- Adjusted the Discover and Feed pages to use array syntax for query keys.

+++ frontend/src/pages/AddEdit/EditPhoto.tsx
@@ -6,6 +6,8 @@ import { photoService } from '../../api/photoService';
 import { useAuth } from '../../hooks/useAuth';
 import { LoadingModal } from '../../components/common/LoadingModal';
 import LoadingSpinner from '../../components/common/LoadingSpinner';
+import axios from 'axios';
+import type { ApiResponse } from '../../types/api.types';
 
 const EditPhoto = () => {
     const { photoId } = useParams();
@@ -24,7 +26,15 @@ const EditPhoto = () => {
                 const photoData = response.data.data;
                 setPhotoData(photoData);
             } catch (error) {
-                const errorMessage = error.response?.data?.message || error.message || "Đã có lỗi xảy ra. Vui lòng thử lại!";
+                let errorMessage = "Đã có lỗi xảy ra. Vui lòng thử lại!";
+
+                if (axios.isAxiosError<ApiResponse<string>>(error)) {
+                    errorMessage = error.response?.data?.message || error.message;
+                }
+
+                else if (error instanceof Error) {
+                    errorMessage = error.message;
+                }
                 toast.error(errorMessage);
                 setHasError(true);
             } finally {
@@ -39,11 +49,20 @@ const EditPhoto = () => {
         console.log(formData)
         try {
             setIsUploading(true);
+            if (!photoId) return;
             await photoService.editPhoto(photoId, formData, isAdmin);
             toast.success("Cập nhật thành công!");
-            setTimeout(() => navigate('/'), 2000);
+
         } catch (error) {
-            const errorMessage = error.response?.data?.message || error.message || "Đã có lỗi xảy ra. Vui lòng thử lại!";
+            let errorMessage = "Đã có lỗi xảy ra. Vui lòng thử lại!";
+
+            if (axios.isAxiosError<ApiResponse<string>>(error)) {
+                errorMessage = error.response?.data?.message || error.message;
+            }
+
+            else if (error instanceof Error) {
+                errorMessage = error.message;
+            }
             toast.error(errorMessage);
         } finally {
             setIsUploading(false);
@@ -53,11 +72,20 @@ const EditPhoto = () => {
     const handleDelete = async () => {
         try {
             setIsUploading(true);
+            if (!photoId) return;
             await photoService.deletePhoto(photoId, isAdmin);
-            toast.success("Xóa ảnh thành công!");
-            setTimeout(() => navigate('/'), 2000);
+            toast.success("Xóa ảnh thành công! ");
+
         } catch (error) {
-            const errorMessage = error.response?.data?.message || error.message || "Đã có lỗi xảy ra. Vui lòng thử lại!";
+            let errorMessage = "Đã có lỗi xảy ra. Vui lòng thử lại!";
+
+            if (axios.isAxiosError<ApiResponse<string>>(error)) {
+                errorMessage = error.response?.data?.message || error.message;
+            }
+
+            else if (error instanceof Error) {
+                errorMessage = error.message;
+            }
             toast.error(errorMessage);
             throw error;
         } finally {
